requirements.txt
groq>=0.9
python-dotenv>=1.0
